requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
tomli-w>=1.0.0
tomllib>=2.0.1 ; python_version < "3.11"
